Adds types to WorkoutTemplateBuilder
